requirements.txt
asgiref==3.6.0
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
cloudinary==1.32.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==40.0.2
defusedxml==0.7.1
Django==3.2.16
django-autoslug==1.9.9
django-ckeditor==6.3.2
django-cloudinary-storage==0.3.0
django-cors-headers==3.13.0
django-environ==0.9.0
django-js-asset==2.0.0
django-leaflet==0.29.0
django-storages==1.13.1
django-templated-mail==1.1.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
djoser==2.2.0
drf-yasg==1.21.5
GDAL @ file:///C:/Users/topga/Downloads/GDAL-3.4.3-cp311-cp311-win_amd64.whl
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
MarkupSafe==2.1.2
oauthlib==3.2.2
packaging==23.1
Pillow==9.3.0
psycopg2==2.9.5
pycparser==2.21
PyJWT==2.6.0
python-slugify==8.0.1
python3-openid==3.2.0
pytz==2023.3
requests==2.29.0
requests-oauthlib==1.3.1
ruamel.yaml==0.17.22
six==1.16.0
social-auth-app-django==5.2.0
social-auth-core==4.4.2
sqlparse==0.4.4
text-unidecode==1.3
uritemplate==4.1.1
urllib3==1.26.15
